BlueToothPairing
Bluetooth class, which describes general characteristics and
capabilities of a device. For example, Generally when we connect to a
Bluetooth Device like Computer, Laptop, Bluetooth Headset and Bluetooth
other peripherals.And the Bluetooth class will specify the general
device type such as a phone, a computer, or headset, and whether it's
capable of services such as audio or telephony.

And today I am adding an other basic tutorial to get List of Paired
Bluetooth Devices and get the Major Class (Headset, Computer or other
Peripheral) of the Paired Bluetooth Devices